Refactor FieldLabelContent to LabelContent
