+++ src/components/designsystemet/TextArea/TextArea.tsx
@@ -1,5 +1,5 @@
 import { Textfield as DsTextField } from '@digdir/designsystemet-react';
-import { FieldLabelContent } from '~/main';
+import { LabelContent } from '~/main';
 import { InputSize, inputSizeClass } from '~/utils/input/input';
 
 export const Textfield = null;
@@ -26,7 +26,7 @@ export const TextArea = ({ size = 'full', className, ...props }: TextAreaProps) 
   return (
     <DsTextField
       className={`${className} ${inputSizeClass(size)}`}
-      label={<FieldLabelContent text={props.label} required={props.required} optional={props.optional} />}
+      label={<LabelContent text={props.label} required={props.required} optional={props.optional} />}
       description={props.description}
       readOnly={props.readOnly}
       placeholder={props.placeholder}
