--- src/components/kystverket/ErrorLabel/errorLabel.module.css
@@ -1,7 +1,0 @@
-.text {
-  color: var(--ds-color-danger-text-subtle);
-}
-.icon {
-  color: var(--ds-color-danger-text-subtle);
-  font-size: 24px;
-}
